Fix review round 9: Layer A click coordinates were dropped or fractional
Two related coordinate bugs in the F4 oracle tap's Layer A routes:

- debug.recordOutcome read where.x/y with Lua's tointeger, which
  refuses any non-integral number outright — a fractional click
  position (sub-pixel screen coords, or the playtest harness's
  deliberately non-integral injections) silently dropped the WHOLE
  `where` field, not just the fraction. Switched to tonumber via a new
  getSubDouble helper, and widened ActionOutcome's aoWhereX/aoWhereY
  from Maybe Int to Maybe Double end to end (the Lua pusher, and every
  Haskell producer's Int tile coordinates now wrapped with
  fromIntegral: Cursor/Common.hs, Cursor/Plant.hs).

- Several engine-owned click routes recorded no coordinates at all,
  regardless of fraction: Engine.Input.Thread's swallowed/no-handler
  routes (recordRouteOutcome) hard-coded Nothing/Nothing instead of
  threading the real press position, and the UI-widget route
  (LuaUIClickEvent/LuaUIRightClickEvent, dispatched in
  Engine.Scripting.Lua.Thread.Dispatch) had no coordinate field on the
  message at all. Both now carry the click's real window-space
  position, matching the convention scripts/init_mouse.lua's
  recordClick already uses for the game-chain routes.

Regression coverage: action_outcome_probe.py gains an explicit
fractional-coordinate round-trip test (the exact review repro), and
the windowed action_outcome_layer_a_check.py now asserts `where`
matches the real click position (converted from framebuffer to
window space) on the widget, no-right-click-handler, and deadclick
routes it already exercises.

+++ src/Engine/ActionOutcome.hs
@@ -36,8 +36,14 @@ data ActionOutcome = ActionOutcome
       -- ^ which action/verb, e.g. "till.designate", "input.click"
     , aoOutcome   ∷ !Text
       -- ^ "accepted" | "rejected" | "partial" | "noop" | "deadclick"
-    , aoWhereX    ∷ !(Maybe Int)
-    , aoWhereY    ∷ !(Maybe Int)
+    , aoWhereX    ∷ !(Maybe Double)
+      -- ^ Double, not Int: world-side producers always pass whole tile
+      --   coordinates, but Layer A's screen-space clicks are frequently
+      --   fractional (sub-pixel positions, and the playtest harness
+      --   deliberately injects non-integral coordinates) — an Int field
+      --   would silently drop the whole `where` on every such click
+      --   (review round 9).
+    , aoWhereY    ∷ !(Maybe Double)
     , aoTarget    ∷ !(Maybe Word32)
     , aoRequested ∷ !(Maybe Int)
       -- ^ for "partial": tiles/units the sweep covered before filtering
